+++ common/mq/tryMakeFriends.go
@@ -9,11 +9,11 @@ import (
 const TypeTryMakeFriends = "friend:make"
 
 type TryMakeFriendsPayload struct {
-	UserAId int32
-	UserBId int32
+	UserAId int64
+	UserBId int64
 }
 
-func NewTryMakeFriendsTask(userAId int32, userBId int32) (*asynq.Task, error) {
+func NewTryMakeFriendsTask(userAId int64, userBId int64) (*asynq.Task, error) {
 	payload, err := json.Marshal(TryMakeFriendsPayload{UserAId: userAId, UserBId: userBId})
 	if err != nil {
 		return nil, errors.Wrapf(err, "failed to marshal payload for task %q", TypeTryMakeFriends)
